feat(rice-warehouse): add sales-to-outbound flow, non-paddy purchasing, rice reports, and putaway scoring

+++ Backend.Application/DTOs/Locations/LocationDetailDto.cs
@@ -11,10 +11,10 @@ public class LocationDetailDto
     public string? ShelfRow { get; set; }
     public string? ShelfLevel { get; set; }
     public string? SlotCode { get; set; }
-    public int? MaxCapacity { get; set; }
+    public decimal? MaxCapacity { get; set; }
     public string? Description { get; set; }
     public bool IsActive { get; set; }
-    public int CurrentOccupancy { get; set; }
+    public decimal CurrentOccupancy { get; set; }
     public int? AllowedCategoryId { get; set; }
     public string? AllowedCategoryName { get; set; }
     public int Priority { get; set; }
